+++ torchao/quantization/pt2e/quantizer/__init__.py
@@ -1,4 +1,6 @@
 from .composable_quantizer import ComposableQuantizer
+from .duplicate_dq_pass import DuplicateDQPass
+from .port_metadata_pass import PortNodeMetaForQDQ
 from .quantizer import (
     DerivedQuantizationSpec,
     EdgeOrNode,
@@ -10,6 +12,8 @@ from .quantizer import (
     SharedQuantizationSpec,
 )
 from .utils import (
+    OperatorConfig,
+    OperatorPatternType,
     QuantizationConfig,
     _annotate_input_qspec_map,
     _annotate_output_qspec,
@@ -38,8 +42,12 @@ __all__ = [
     "_get_module_name_filter",
     "_is_valid_annotation",
     "QuantizationConfig",
+    "OperatorPatternType",
+    "OperatorConfig",
     "get_input_act_qspec",
     "get_output_act_qspec",
     "get_weight_qspec",
     "get_bias_qspec",
+    "DuplicateDQPass",
+    "PortNodeMetaForQDQ",
 ]
